+++ branches/zmianaZaleznosci/TGUI/src/TGUI/ListBox.cpp
@@ -229,7 +229,7 @@ namespace tgui
                     m_Scroll->setVerticalScroll(true);
                     m_Scroll->setSize(m_Scroll->getSize().x, static_cast<float>(m_Size.y));
                     m_Scroll->setLowValue(m_Size.y);
-                    m_Scroll->setMaximum(m_Items.size() * m_ItemHeight);
+                    m_Scroll->setMaximum(static_cast<unsigned int>(m_Items.size())* m_ItemHeight);
                 }
             }
             else
@@ -411,10 +411,10 @@ namespace tgui
 
             // If there is a scrollbar then tell it that another item was added
             if (m_Scroll != nullptr)
-                m_Scroll->setMaximum(m_Items.size() * m_ItemHeight);
+				m_Scroll->setMaximum(static_cast<unsigned int>(m_Items.size()) * m_ItemHeight);
 
             // Return the item index
-            return m_Items.size() - 1;
+			return static_cast<int>(m_Items.size()) - 1;
         }
         else // The item limit was reached
             return -1;
@@ -489,7 +489,7 @@ namespace tgui
 
         // If there is a scrollbar then tell it that an item was removed
         if (m_Scroll != nullptr)
-            m_Scroll->setMaximum(m_Items.size() * m_ItemHeight);
+			m_Scroll->setMaximum(static_cast<unsigned int>(m_Items.size()) * m_ItemHeight);
 
         // Check if the selected item should change
         if (m_SelectedItem == static_cast<int>(index))
@@ -520,7 +520,7 @@ namespace tgui
 
                 // If there is a scrollbar then tell it that an item was removed
                 if (m_Scroll != nullptr)
-                    m_Scroll->setMaximum(m_Items.size() * m_ItemHeight);
+					m_Scroll->setMaximum(static_cast<unsigned int>(m_Items.size()) * m_ItemHeight);
 
                 return true;
             }
@@ -632,7 +632,7 @@ namespace tgui
             m_Scroll->setVerticalScroll(true);
             m_Scroll->setSize(m_Scroll->getSize().x, static_cast<float>(m_Size.y));
             m_Scroll->setLowValue(m_Size.y);
-            m_Scroll->setMaximum(m_Items.size() * m_ItemHeight);
+			m_Scroll->setMaximum(static_cast<unsigned int>(m_Items.size()) * m_ItemHeight);
 
             return true;
         }
@@ -685,7 +685,7 @@ namespace tgui
         else // There is a scrollbar
         {
             // Set the maximum of the scrollbar
-            m_Scroll->setMaximum(m_Items.size() * m_ItemHeight);
+			m_Scroll->setMaximum(static_cast<unsigned int>(m_Items.size()) * m_ItemHeight);
         }
     }
 
@@ -711,7 +711,7 @@ namespace tgui
 
             // If there is a scrollbar then tell it that the number of items was changed
             if (m_Scroll != nullptr)
-                m_Scroll->setMaximum(m_Items.size() * m_ItemHeight);
+				m_Scroll->setMaximum(static_cast<unsigned int>(m_Items.size()) * m_ItemHeight);
         }
     }
